countdown: cleanup

+++ src/main/java/com/abaan404/boatrace/game/race/RaceWidgets.java
@@ -69,7 +69,7 @@ public class RaceWidgets {
             return;
         }
 
-        long countdown = stageManager.countdown.getCountdown();
+        int countdown = stageManager.countdown.getCountdown();
         if (countdown <= 0) {
             this.shownGo = true;
         }
